Add PKCS#11 URI handling support in lib/p11uri
Since p11-kit isn't present on all platforms we can't rely on linking to
it dynamically. To import it wholesale as 'external code' like we do
zlib and sqlite seems to make little sense since we only need a small
part of it.

Instead we only include necessary part of the PKCS#11 URI stuff from
p11-kit, with mangling names to our convention and export symbols needed
by other part of NSS.