Make NSSTrustDomain_FindTokenByName() take a ref on the token it returns
AFAICT without this, there's nothing that guarantees that 'tok' still
actually exists, from the moment we drop the locks and return it.

This function isn't exported, and is used precisely once in the entire
code base. So make it take a reference on the token, and make its only
caller release that reference.